+++ jgiven-core/src/main/java/com/tngtech/jgiven/annotation/ScenarioDescription.java
@@ -7,13 +7,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Can be used to annotated test methods to give a 
+ * Can be used to annotated test methods to give a
  * scenario a complex description that cannot be
- * directly represented by a Java method name. 
+ * directly represented by a Java method name.
+ *
+ * @deprecated just use {@link @Description} instead, will be removed with v0.5
  */
 @Retention( RetentionPolicy.RUNTIME )
 @Target( ElementType.METHOD )
 @Documented
+@Deprecated
 public @interface ScenarioDescription {
     String value();
 }